+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Added
 
 - `moonstone/VideoPlayer.MediaControls` property `playerNode`
+## [Unreleased]
+
+### Fixed
+- `moonstone/Scroller` ordering of logic for Scroller focus to check focus possibilities first then go to fallback at the top of the container
 
 ## [2.0.0-beta.5] - 2018-05-29